Added ARTEMIS_VERSION preprocessor definition
APP_VER and VERSIONSTR are defined in the makefile. Changing these two values with affect the PARAM.SFO and ARTEMIS_VERSION preprocessor definition.

Learning from Artemis PS3, I'm keeping the number of instances of the application version to a minimum. This way, updating the version is done within one file instead of many.