Added footer to WindowFrame class
Implemented footer in classes that use WindowFrame

Moved 'wstringEmpty' from Locale to 'WSTRING_EMPTY' in Globals.
Most methods (in this project) that take a string just hold the
reference instead of copying it. Passing 'L""' gives a temporary empty
string that will be overwritten in the stack at a later point causing
the reference to become invalid. Having this global constant bypasses
this issue.